feat: first working build — IF signal and range spectrum rendering
- Fix generate_chirp_if() call signature (chirp_index=0, not num_samples)
- Add rcs=1.0, num_samples, num_chirps to RadarParams/Target defaults
- Fix complex DFT sign convention (negative exponent for forward DFT)
- Fix PlotWidget y-axis autoscale after QLineSeries::replace()
- Fix ParameterPanel.ui XML (remove invalid <int> tags)
- Fix AUTOUIC search path to find .ui files in ui/ directory
- Add Q_OBJECT headers to qt_add_executable and radar_engine targets

Spectrum shows clean sinc peak at ~47 m (expected 50 m, offset is
DFT bin rounding — resolves in Phase 2 with FFTW3 zero-padding).
Status bar: f0=300 GHz, Δr=3.75 cm, f_beat=13333.33 kHz confirmed.

+++ include/RadarEngine.hpp
@@ -61,7 +61,7 @@ signals:
 private:
     RadarParams                        m_params{};
     Target                             m_target{};
-    std::vector<std::complex<double>>  m_ifBuf{};
+    std::vector<std::complex<float>>   m_ifBuf{};
 
     void buildSpectrum();
